Rename COMP_TYPE to MATH_PREC, add templated setScalarVector/getScalarVector, use int32 block_size
- Rename HIPDNN_ATTR_BLOCK_SCALE_DEQUANTIZE_COMP_TYPE_EXT to MATH_PREC_EXT
- Add templated setScalarVector/getScalarVector to DescriptorAttributeUtils
- Store block_size as int32 natively (no int64 conversion)
- Template setDescriptorAttrVec in DescriptorHelpers
- Refactor tests: setAllAttributesExcept helper, parameterized finalize/getTensor/query tests
- Add is_negative_scale coverage throughout tests
- Set is_negative_scale=true in integration test

+++ projects/hipdnn/backend/src/descriptors/BlockScaleDequantizeOperationDescriptor.hpp
@@ -71,8 +71,8 @@ private:
     std::shared_ptr<TensorDescriptor> _scaleDesc;
     std::shared_ptr<TensorDescriptor> _yDesc;
 
-    // Block size stored as int64 for the C-API (converted to int32 for FBS in buildNode)
-    std::vector<int64_t> _blockSize;
+    // Block size stored as int32 (matches the FBS schema type)
+    std::vector<int32_t> _blockSize;
 
     // Compute data type for this operation (stored at node level in graph)
     hipdnn_data_sdk::data_objects::DataType _computeDataType
